Move batch premake to scripts directory, organize source into other dirs, fix compiler warning for void ptr truncation

+++ Skye/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -65,7 +65,7 @@ namespace Skye {
 				ShaderDataTypeToOpenGLBaseType(element.Type),
 				element.Normalized ? GL_TRUE : GL_FALSE,
 				layout.GetStride(),
-				(const void*)element.Offset
+				(const void*)(size_t)(element.Offset)
 			);
 			index++;
 		}
